+++ example/lib/main.dart
@@ -3,8 +3,6 @@ import 'dart:async';
 import 'package:arcane/arcane.dart';
 import 'package:example/model/note.dart';
 import 'package:example/screen/home.dart';
-import 'package:example/screen/note_view.dart';
-import 'package:example/screen/notes.dart';
 
 bool v = false;
 String? vv;
@@ -26,8 +24,6 @@ class ExampleArcaneApp extends StatelessWidget {
   Widget build(BuildContext context) => ArcaneApp(
         arcaneRoutes: [
           HomeScreen(),
-          NotesScreen(),
-          NoteScreen(),
         ],
         theme: ArcaneTheme(
             themeMode: ThemeMode.system,